--- dhis-2/dhis-api/src/main/java/org/hisp/dhis/icon/CustomIcon.java
@@ -1,96 +1,0 @@
-/*
- * Copyright (c) 2004-2023, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.icon;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hisp.dhis.common.BaseIdentifiableObject;
-import org.hisp.dhis.common.DxfNamespaces;
-import org.hisp.dhis.fileresource.FileResource;
-
-/** Custom icons are uploaded by users and can be modified and deleted. */
-@Getter
-@Setter
-@NoArgsConstructor
-@JacksonXmlRootElement(localName = "customIcon", namespace = DxfNamespaces.DXF_2_0)
-public class CustomIcon extends BaseIdentifiableObject {
-
-  public static final String DEFAULT_FIELDS_PARAM = "key,description,keywords,href";
-
-  @JsonProperty private String key;
-
-  @JsonProperty private String description;
-
-  @JsonProperty private Set<String> keywords = new HashSet<>();
-
-  @JsonProperty private Boolean custom;
-
-  @JsonProperty
-  @JsonSerialize(as = BaseIdentifiableObject.class)
-  private FileResource fileResource;
-
-  public CustomIcon(
-      String key,
-      String description,
-      Set<String> keywords,
-      boolean custom,
-      FileResource fileResource) {
-    this.key = key;
-    this.description = description;
-    this.keywords = keywords;
-    this.custom = custom;
-    this.fileResource = fileResource;
-    this.setAutoFields();
-  }
-
-  @Override
-  public boolean equals(Object obj) {
-    if (this == obj) {
-      return true;
-    }
-
-    if (obj == null || getClass() != obj.getClass()) {
-      return false;
-    }
-
-    CustomIcon other = (CustomIcon) obj;
-    return Objects.equals(uid, other.uid) && Objects.equals(key, other.key);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(uid, key);
-  }
-}
