+++ dhis-2/dhis-services/dhis-service-schema/src/main/java/org/hisp/dhis/schema/descriptors/IconSchemaDescriptor.java
@@ -27,7 +27,7 @@
  */
 package org.hisp.dhis.schema.descriptors;
 
-import org.hisp.dhis.icon.CustomIcon;
+import org.hisp.dhis.icon.Icon;
 import org.hisp.dhis.schema.Schema;
 import org.hisp.dhis.schema.SchemaDescriptor;
 
@@ -43,7 +43,7 @@ public class IconSchemaDescriptor implements SchemaDescriptor {
 
   @Override
   public Schema getSchema() {
-    Schema schema = new Schema(CustomIcon.class, SINGULAR, PLURAL);
+    Schema schema = new Schema(Icon.class, SINGULAR, PLURAL);
     schema.setRelativeApiEndpoint(API_ENDPOINT);
 
     return schema;
